// src/http/models/getPartners200Item.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * onec
 * OpenAPI spec version: 1.0.0
 */

export type GetPartners200Item = {
  id: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  cpfOrCnpj: string | null;
  /** @nullable */
  city: string | null;
  /** @nullable */
  state: string | null;
  /** @nullable */
  commission: number | null;
  /** @nullable */
  portal: string | null;
  /** @nullable */
  channelHead: string | null;
  /** @nullable */
  regional: string | null;
  /** @nullable */
  coordinator: string | null;
  /** @nullable */
  agent: string | null;
  /** @nullable */
  indicator: string | null;
  /** @nullable */
  contract: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  responsible: string | null;
};